+++ templates/parts/menu/menu-primary.php
@@ -12,7 +12,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.3.0
+ * @version  1.3.2
  */
 
 namespace WebManDesign\Michelle;
@@ -24,11 +24,7 @@ $is_mobile_nav_enabled = Customize\Mod::get( 'navigation_mobile' );
 
 ?>
 
-<nav
-	id="site-navigation"
-	class="main-navigation"
-	aria-label="<?php echo esc_attr_x( 'Main menu', 'Navigational menu.', 'michelle' ); ?>"
->
+<nav id="site-navigation" class="main-navigation" aria-label="<?php echo esc_attr_x( 'Main menu', 'Navigational menu.', 'michelle' ); ?>">
 
 	<?php